#5 - install new react-ui kit
