+++ package.json
@@ -9,6 +9,7 @@
   "dependencies": {
     "react": "16.0.0-alpha.12",
     "react-native": "0.46.4",
+    "react-native-elements": "^0.15.0",
     "react-native-gifted-messenger": "^0.1.4",
     "react-native-value-store": "^0.1.0",
     "react-native-vector-icons": "^4.2.0",
