--- src/data/contacts.json
@@ -1,10 +1,0 @@
-[
-  {key: 'Albert'},
-  {key: 'Bennt'},
-  {key: 'Cidar'},
-  {key: 'Dustin'},
-  {key: 'Evelyn'},
-  {key: 'Gary'},
-  {key: 'Hector'},
-  {key: 'Ivan'},
-]
